packages/rag/query/requirements.txt
openai
minio
farm-haystack
setuptools
nltk
langdetect
